+++ lib/SIL/Utils/CMakeLists.txt
@@ -1,5 +1,6 @@
 target_sources(swiftSIL PRIVATE
   BasicBlockUtils.cpp
+  BitDataflow.cpp
   DebugUtils.cpp
   Dominance.cpp
   DynamicCasts.cpp
@@ -7,6 +8,7 @@ target_sources(swiftSIL PRIVATE
   InstructionUtils.cpp
   LoopInfo.cpp
   MemAccessUtils.cpp
+  MemoryLocations.cpp
   OptimizationRemark.cpp
   OwnershipUtils.cpp
   PrettyStackTrace.cpp
