+++ lib/SIL/Verifier/CMakeLists.txt
@@ -1,7 +1,7 @@
 target_sources(swiftSIL PRIVATE
   LoadBorrowImmutabilityChecker.cpp
   LinearLifetimeChecker.cpp
-  MemoryLifetime.cpp
+  MemoryLifetimeVerifier.cpp
   ReborrowVerifier.cpp
   SILOwnershipVerifier.cpp
   SILVerifier.cpp)
