+++ lib/SILOptimizer/Mandatory/OptimizeHopToExecutor.cpp
@@ -14,7 +14,7 @@
 #include "swift/SIL/SILBuilder.h"
 #include "swift/SIL/SILFunction.h"
 #include "swift/SIL/ApplySite.h"
-#include "swift/SIL/MemoryLifetime.h"
+#include "swift/SIL/MemoryLocations.h"
 #include "swift/SILOptimizer/PassManager/Transforms.h"
 #include "swift/SIL/MemAccessUtils.h"
